Moving LA Port 2 to LA Port
